Can now write to PriceMap.txt
