+++ src/fluttersnackmap/assets/VendingInformation.txt
@@ -10,7 +10,7 @@ PS 1st floor near bathrooms 1,Working,Accepts Cash/Crimson Card/Credit and Debit
 , ,CV 2nd floor near stairs 2,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Drink,Coca-Cola,1.75,Coca-Cola Cherry,1.75,Coca-Cola Diet,1.75,Sprite,1.75,Mello Yello,1.75,Fanta,1.75
 , ,Library 1st floor near office,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Drink,Coca-Cola,1.75,Coca-Cola Zero,1.75,Coca-Cola Diet,1.75,Coca-Cola Cherry,1.75,Pibb Xtra,1.75,Dasani Water,1.75
 , ,HH 2nd floor near breezeway 1,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Drink,Coca-Cola,1.75,Coca-Cola Zero,1.75,Coca-Cola Cherry,1.75,Coca-Cola Diet,1.75,Sprite,1.75,Mello Yello,1.75
-, ,HH 2nd floor near breezeway 2,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Snack,Grippos BBQ,1.50,Oatmeal Cream Pie,1.25,Reese's Cups,2.00,Reese's Pieces,2.00,Baby Ruth,2.00,Belvita Crunchy,2.00,Twix,2.00,3 Musketeers,2.00,Snickers,2.00,M&M's Peanuts,2.00,Reese's Take 5,2.00,Butter Finger,2.00,Nature Valley Crunchy,2.00,Lance Peanut Butter Crackers,1.25, Salted Peanuts,2.00,Sweet 'N Spicy Mix,200,Oreo,200,Pork and Beef Sticks,200,Skittles,200,Starburst Mini,200,Cliff Chocolate Chip Cookies,2.00,Rice Krispie Treat,200,Pop-Tart Strawberry,2.25,Mamba,2.25,Chocolate Chip Cookies,2.25,5 Spearmint,2.00
+, ,HH 2nd floor near breezeway 2,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Snack,Grippos BBQ,1.50,Oatmeal Cream Pie,1.25,Reese's Cups,2.00,Reese's Pieces,2.00,Baby Ruth,2.00,Belvita Crunchy,2.00,Twix,2.00,3 Musketeers,2.00,Snickers,2.00,M&M's Peanuts,2.00,Reese's Take 5,2.00,Butter Finger,2.00,Nature Valley Crunchy,2.00,Lance Peanut Butter Crackers,1.25, Salted Peanuts,2.00,Sweet 'N Spicy Mix,2.00,Oreo,2.00,Pork and Beef Sticks,2.00,Skittles,2.00,Starburst Mini,2.00,Cliff Chocolate Chip Cookies,2.00,Rice Krispie Treat,2.00,Pop-Tart Strawberry,2.25,Mamba,2.25,Chocolate Chip Cookies,2.25,5 Spearmint,2.00
 , ,HH 3rd floor in common area 1,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Drink,Coca-Cola,1.75,Coca-Cola Zero,1.75,Coca-Cola Diet,1.75,Coca-Cola Cherry,1.75,Sprite,1.75,Pibb Xtra,1.75
 , ,HH 3rd floor in common area 2,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Snack,Pork Rinds BBQ,1.50,Reese's Cups,2.00,Reese's Pieces,2.00,Baby Ruth,2.00,Skittles,2.00,3 Musketeers,2.00,Nature Valley Crunchy,2.00,Lance Grilled Cheese Crackers,1.25,Mamba,1.25,Nutter Butter,2.00,Rice Krispie Treat,2.00,Life Savers,1.00
 , ,UCS Ground floor near book store,Working,Accepts Cash/Crimson Card/Credit and Debit Card/Apple Pay/ Samsung Pay/Google Pay,Drink,Coca-Cola,1.75,Coca-Cola Zero,1.75,Coca-Cola Cherry,1.75,Coca-Cola Diet,1.75,Sprite,1.75,Dasani Water,1.75
