+++ src/fluttersnackmap/pubspec.yaml
@@ -61,6 +61,7 @@ flutter:
 
   assets:
    - assets/VendingInformation.txt
+   - assets/PriceMap.txt
 
   # To add assets to your application, add an assets section, like this:
   # assets:
