+++ src/fluttersnackmap/test/read_file_test.dart
@@ -1,5 +1,8 @@
+import 'dart:io';
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fluttersnackmap/read_and_write.dart';
+import 'package:fluttersnackmap/search_by_price.dart';
 
 
 
@@ -25,6 +28,15 @@ void main() {
 
   
   });
+  test('Write to PriceMap.txt works',(){
+    Map<int,Map<String,String>> allVendingMap = readFile('assets/VendingInformation.txt');
+    Map<String, List<String>?> temp = SearchByPrice().sortByPrice(allVendingMap);
+    writePriceMap(temp);
+    var myFile = File('assets/PriceMap.txt'); 
+    var actual = myFile.readAsLinesSync().length;
+    expect(actual,89);
+  });
 
   
 }
+
